feat(zz_fmt): M6 CLI polish + M7 LSP integration
Wire CLI execution and LSP formatting through zz_fmt:
- zz fmt uses rayon parallel formatting via format_sources_parallel
- gitignore-aware file discovery via zz_fmt::discover
- --check mode reads-only, no disk writes
- --stdin mode reads from stdin, outputs to stdout
- -c short flag normalized in parse_path_and_flags
- LSP routes document formatting to zz_fmt::format_source
  (full verification pipeline: lex → parse → format → verify)
- Add 11 CLI e2e tests (check, stdin, in-place, directory)
- Fix ~80 clippy warnings across all crates
- 583 tests pass, 0 clippy warnings

+++ crates/zz_frontend/src/levenshtein.rs
@@ -40,7 +40,7 @@ pub fn levenshtein(a: &str, b: &str) -> u32 {
     let shorter_len = alen.min(blen);
     let threshold = max_distance_by_len(shorter_len);
     if len_diff > threshold {
-        return len_diff as u32;
+        return len_diff;
     }
 
     // Single-row DP.
